pyproject.toml
[project]
name = "casamento"
version = "0.1.0"
description = ""
authors = [
    {name = "hallan-n",email = "hallanneves01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "nicegui (>=2.20.0,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "cryptography (>=45.0.4,<46.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

